# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Campus Skill Share API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface AuthUser {
  id: string;
  /** @nullable */
  email: string | null;
  /** @nullable */
  firstName: string | null;
  /** @nullable */
  lastName: string | null;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface AuthUserEnvelope {
  user: AuthUser | null;
}

export interface MobileTokenExchangeRequest {
  /** @minLength 1 */
  code: string;
  /** @minLength 1 */
  code_verifier: string;
  /** @minLength 1 */
  redirect_uri: string;
  /** @minLength 1 */
  state: string;
  /** @minLength 1 */
  nonce?: string;
}

export interface MobileTokenExchangeSuccess {
  token: string;
}

export const LogoutSuccessValue = {
  success: true,
} as const;
export type LogoutSuccess = typeof LogoutSuccessValue;

export interface ErrorEnvelope {
  error: string;
}

export type UserProfileRole = typeof UserProfileRole[keyof typeof UserProfileRole];


export const UserProfileRole = {
  student: 'student',
  admin: 'admin',
} as const;

export interface UserProfile {
  id: string;
  /** @nullable */
  email: string | null;
  displayName: string;
  role: UserProfileRole;
  /** @nullable */
  profileImageUrl: string | null;
  createdAt: string;
}

export interface UserProfileUpdate {
  /** @minLength 2 */
  displayName?: string;
}

export interface PostAuthor {
  id: string;
  displayName: string;
  /** @nullable */
  profileImageUrl: string | null;
}

export interface Post {
  id: string;
  title: string;
  category: string;
  description: string;
  /** @nullable */
  availability?: string | null;
  /** @nullable */
  priceRate?: string | null;
  createdAt: string;
  author: PostAuthor;
}

export type PostInputCategory = typeof PostInputCategory[keyof typeof PostInputCategory];


export const PostInputCategory = {
  Tutoring: 'Tutoring',
  Design: 'Design',
  Music: 'Music',
  Tech: 'Tech',
  Language: 'Language',
  Other: 'Other',
} as const;

export interface PostInput {
  /**
     * @minLength 3
     * @maxLength 150
     */
  title: string;
  category: PostInputCategory;
  /**
     * @minLength 20
     * @maxLength 500
     */
  description: string;
  /** @maxLength 200 */
  availability?: string;
  /** @maxLength 100 */
  priceRate?: string;
}

export interface PostListResponse {
  posts: Post[];
  total: number;
}

export interface CategoryStat {
  category: string;
  count: number;
}

export interface PostStatsResponse {
  categories: CategoryStat[];
  total: number;
}

export interface Message {
  id: string;
  senderId: string;
  receiverId: string;
  content: string;
  createdAt: string;
}

export interface MessageInput {
  receiverId: string;
  /**
     * @minLength 1
     * @maxLength 2000
     */
  content: string;
}

export interface MessageListResponse {
  messages: Message[];
}

export interface ConversationSummary {
  otherUser: PostAuthor;
  lastMessage: Message;
  unreadCount: number;
}

export interface ConversationListResponse {
  conversations: ConversationSummary[];
}

export interface AdminStats {
  totalUsers: number;
  totalPosts: number;
  postsByCategory: CategoryStat[];
}

/**
 * Opaque session token — `Bearer <sid>`.
 */
export type AuthorizationSessionHeaderParameter = string;

export type BeginBrowserLoginParams = {
returnTo?: string;
};

export type HandleBrowserLoginCallbackParams = {
code?: string;
state?: string;
iss?: string;
};

export type LogoutBrowserSessionParams = {
returnTo?: string;
};

export type ListPostsParams = {
search?: string;
category?: string;
limit?: number;
offset?: number;
};

export type AdminListPostsParams = {
search?: string;
category?: string;
};