# lib/api-zod/src/generated/types/authUserEnvelope.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Campus Skill Share API
 * OpenAPI spec version: 0.1.0
 */
import type { AuthUser } from './authUser';

export interface AuthUserEnvelope {
  user: AuthUser | null;
}